+++ labelbox/client.py
@@ -611,7 +611,7 @@ class Client:
 
         Attribute values are passed as keyword arguments.
 
-        >>> project = client.create_project(name="<project_name>", description="<project_description>")
+        >>> project = client.create_project(name="<project_name>", description="<project_description>", queue_mode=QueueMode.Batch)
 
         Args:
             **kwargs: Keyword arguments with Project attribute values.